fixing unit test failing
